@@ -3,15 +3,15 @@ apply plugin: 'com.android.application'
 apply from: "$rootDir/versions.gradle"
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion "25.0.0"
+    compileSdkVersion 28
+    buildToolsVersion "28.0.3"
     defaultConfig {
         applicationId "agency.tango.materialintro"
-        minSdkVersion 15
-        targetSdkVersion 25
+        minSdkVersion 19
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
         release {
@@ -22,11 +22,8 @@ android {
 }
 
 dependencies {
-    compile fileTree(include: ['*.jar'], dir: 'libs')
-    androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
-        exclude group: 'com.android.support', module: 'support-annotations'
-    })
-    compile "com.android.support:appcompat-v7:${project.androidSupport}"
-    testCompile 'junit:junit:4.12'
-    compile project(path: ':material-intro-screen')
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
+    implementation project(path: ':material-intro-screen')
+    implementation "androidx.appcompat:appcompat:1.0.2"
+    testImplementation 'junit:junit:4.12'
 }
