@@ -2,8 +2,8 @@ package agency.tango.materialintro;
 
 import android.Manifest;
 import android.os.Bundle;
-import android.support.annotation.FloatRange;
-import android.support.annotation.Nullable;
+import androidx.annotation.FloatRange;
+import androidx.annotation.Nullable;
 import android.view.View;
 import android.widget.Toast;
 
