@@ -1,6 +1,6 @@
 package agency.tango.materialintroscreen.animations;
 
-import android.support.annotation.FloatRange;
+import androidx.annotation.FloatRange;
 import android.view.View;
 
 public interface IViewTranslation {
