@@ -1,6 +1,6 @@
 package agency.tango.materialintroscreen.animations;
 
-import android.support.annotation.AnimRes;
+import androidx.annotation.AnimRes;
 import android.view.View;
 import android.view.animation.Animation;
 import android.view.animation.AnimationUtils;
