@@ -1,13 +1,13 @@
 package agency.tango.materialintroscreen.widgets;
 
 import android.content.Context;
-import android.support.v4.view.CustomViewPager;
-import android.support.v4.view.MotionEventCompat;
+import androidx.core.view.MotionEventCompat;
 import android.util.AttributeSet;
 import android.view.KeyEvent;
 import android.view.MotionEvent;
 
 import agency.tango.materialintroscreen.adapter.SlidesAdapter;
+import androidx.viewpager.widget.CustomViewPager;
 
 public class SwipeableViewPager extends CustomViewPager {
     private float startPos = 0;
