@@ -1,12 +1,12 @@
 package agency.tango.materialintroscreen.adapter;
 
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentStatePagerAdapter;
 import android.view.ViewGroup;
 
 import java.util.ArrayList;
 
 import agency.tango.materialintroscreen.SlideFragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentStatePagerAdapter;
 
 public class SlidesAdapter extends FragmentStatePagerAdapter {
     private ArrayList<SlideFragment> fragments = new ArrayList<>();
